+++ Packages/BrandonUtils/Standalone/Strings/Prettification.cs
@@ -3,6 +3,7 @@ using System.Linq;
 
 using BrandonUtils.Standalone.Collections;
 using BrandonUtils.Standalone.Optional;
+using BrandonUtils.Standalone.Reflection;
 
 using JetBrains.Annotations;