+++ Packages/BrandonUtils/Standalone/Strings/Prettifier.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 
 using BrandonUtils.Standalone.Collections;
+using BrandonUtils.Standalone.Reflection;
 using BrandonUtils.Standalone.Strings.Prettifiers;
 
 using JetBrains.Annotations;