+++ Packages/BrandonUtils/Tests/Standalone/Reflection/ReflectionUtilsTests.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-using BrandonUtils.Standalone;
+using BrandonUtils.Standalone.Reflection;
 
 using NUnit.Framework;